[Bronze II] Title: 시험 감독, Time: 108 ms, Memory: 5928 KB -BaekjoonHub

+++ 백준/Bronze/13458. 시험 감독/시험 감독.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-long long arr[(int)1e6];
+int arr[(int)1e6];
 
 int main() {
   ios_base::sync_with_stdio(false);
